+++ Sources/QuickSearch/String+Types.swift
@@ -3,6 +3,7 @@
 //  QuickSearch
 //
 //  Created by Daniel Saidi on 2023-12-19.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
 import Foundation